requirements.txt
anyio==3.6.1
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
firebase-admin==5.2.0
google-api-core==2.8.1
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.2
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
h11==0.13.0
httplib2==0.20.4
idna==3.3
joblib==1.1.0
msgpack==1.0.4
numpy==1.22.4
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pymongo==3.12.3
pyparsing==3.0.9
requests==2.27.1
rsa==4.8
scikit-surprise==1.1.1
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
python-dotenv==0.20.0
